+++ gatsby/createExifFields.js
